+++ ListDetail/.gitignore
