+++ ListDetail/build.gradle
