+++ ListDetail/src/main/res/drawable/ic_movie_black_24dp.xml
