+++ ListDetail/src/main/res/layout/fragment_item_detail.xml
