+++ ListDetail/src/main/res/layout/fragment_items_list.xml
