+++ ListDetail/src/main/res/mipmap-xxhdpi/ic_launcher_round.png
