+++ ListDetail/src/main/res/values/colors.xml
