+++ ListDetail/src/main/res/values/dimens.xml
