+++ ListDetail/src/main/res/values/styles.xml
