--- MasterDetail/src/main/res/values/strings.xml
@@ -1,14 +1,0 @@
-<!--
-  ~ Copyright (c) Microsoft Corporation. All rights reserved.
-  ~ Licensed under the MIT License.
-  ~
-  -->
-
-<resources>
-    <string name="app_name">Master Detail</string>
-    <string name="rating">Rating:</string>
-    <string name="details">Details:</string>
-    <string name="note">Title:</string>
-    <string name="item_image">Item image</string>
-    <string name="items_list">Items List</string>
-</resources>
